requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Dance==7.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
URLObject==2.4.3
Werkzeug==3.0.1
